Appended "e.g. " to values in example field
Appended the string "e.g. " to the value in the example JSON field such that when the values are displayed in the spreadsheet, they have "e.g. " at the beginning to denote that they are example values. The value in properties[prop]["example"] was type cast to str because some of the example are integers.